backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
litellm==1.55.10
pydantic==2.10.4
sse-starlette==2.2.1
python-dotenv==1.0.1
pyyaml==6.0.3
